# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: SyncHub API
  version: 0.1.0
  description: REST API for SyncHub workspace synchronization.
servers:
  - url: http://localhost:8765
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: User and tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: User and tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Logged out.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/files:
    get:
      tags: [File]
      summary: List files
      parameters:
        - name: parent_id
          in: query
          schema:
            type: string
        - name: page_size
          in: query
          schema:
            type: integer
            maximum: 200
            default: 100
      responses:
        "200":
          description: File list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
  /api/v1/files/by-path:
    get:
      tags: [File]
      summary: Get file by path
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            example: /workspace/readme.txt
      responses:
        "200":
          description: File metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
  /api/v1/files/{id}:
    get:
      tags: [File]
      summary: Get file by id
      parameters:
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: File metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
    patch:
      tags: [File]
      summary: Move or rename file
      parameters:
        - $ref: "#/components/parameters/FileID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                device_id:
                  type: string
                  description: Optional source device id for sync change attribution.
      responses:
        "200":
          description: Updated file metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
    delete:
      tags: [File]
      summary: Delete file
      parameters:
        - $ref: "#/components/parameters/FileID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id:
                  type: string
                  description: Optional source device id for sync change attribution.
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/files/{id}/versions:
    get:
      tags: [File]
      summary: List file versions
      parameters:
        - $ref: "#/components/parameters/FileID"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        "200":
          description: File version history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileVersionListResponse"
  /api/v1/files/{id}/versions/{version}/restore:
    post:
      tags: [File]
      summary: Restore file to a historical version
      parameters:
        - $ref: "#/components/parameters/FileID"
        - name: version
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Restored file metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRestoreResponse"
  /api/v1/files/{id}/versions/{version}/pin:
    post:
      tags: [File]
      summary: Pin file version
      parameters:
        - $ref: "#/components/parameters/FileID"
        - name: version
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Pinned file version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileVersionResponse"
    delete:
      tags: [File]
      summary: Unpin file version
      parameters:
        - $ref: "#/components/parameters/FileID"
        - name: version
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Unpinned file version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileVersionResponse"
  /api/v1/files/directories:
    post:
      tags: [File]
      summary: Create directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  example: /workspace
                device_id:
                  type: string
                  description: Optional source device id for sync change attribution.
      responses:
        "201":
          description: Created directory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
  /api/v1/files/{id}/content:
    get:
      tags: [Download]
      summary: Download file content
      parameters:
        - $ref: "#/components/parameters/FileID"
        - name: Range
          in: header
          schema:
            type: string
            example: bytes=0-1023
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        "200":
          description: File content.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: Partial file content.
  /api/v1/uploads:
    post:
      tags: [Upload]
      summary: Initialize chunk upload
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: Reuses the existing upload session when the same user retries with the same key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInitRequest"
      responses:
        "201":
          description: Upload session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /api/v1/uploads/{id}:
    get:
      tags: [Upload]
      summary: Get upload status
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Upload status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /api/v1/uploads/{id}/chunks/{index}:
    put:
      tags: [Upload]
      summary: Upload chunk
      parameters:
        - $ref: "#/components/parameters/UploadID"
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
        - name: X-Chunk-Sha256
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Uploaded chunk.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkResponse"
  /api/v1/uploads/{id}/commit:
    post:
      tags: [Upload]
      summary: Commit upload session
      parameters:
        - $ref: "#/components/parameters/UploadID"
      responses:
        "200":
          description: Committed upload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitResponse"
  /api/v1/devices:
    post:
      tags: [Sync]
      summary: Register device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, platform]
              properties:
                name:
                  type: string
                  example: work-laptop
                platform:
                  type: string
                  example: windows
      responses:
        "201":
          description: Registered device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
  /api/v1/devices/{id}/heartbeat:
    post:
      tags: [Sync]
      summary: Update device heartbeat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
  /api/v1/sync/changes:
    get:
      tags: [Sync]
      summary: Pull change events
      parameters:
        - name: device_id
          in: query
          required: true
          schema:
            type: string
        - name: after_change_id
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
            maximum: 500
      responses:
        "200":
          description: Change feed page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangeListResponse"
        "410":
          description: Sync cursor expired or outside the available change feed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse"
  /api/v1/sync/ack:
    post:
      tags: [Sync]
      summary: Acknowledge applied changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id, last_applied_change_id]
              properties:
                device_id:
                  type: string
                last_applied_change_id:
                  type: integer
      responses:
        "200":
          description: Updated device cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    APIResponse:
      type: object
      properties:
        code:
          oneOf:
            - type: integer
            - type: string
        message:
          type: string
        data:
          type: object
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  $ref: "#/components/schemas/User"
                tokens:
                  $ref: "#/components/schemas/TokenPair"
    TokenResponse:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TokenPair"
...
    FileResponse:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/FileNode"
    FileListResponse:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/FileNode"
...